+++ base/options.jl
@@ -58,7 +58,7 @@ struct JLOptions
     permalloc_pkgimg::Int8
     heap_size_hint::UInt64
     trace_compile_timing::Int8
-    static_call_graph::Int8
+    trim::Int8
 end
 
 # This runs early in the sysimage != is not defined yet