+++ base/reflection.jl
@@ -1505,11 +1505,11 @@ struct CodegenParams
     use_jlplt::Cint
 
     """
-    If enabled, any attempt to emit a dynamic call will cause codegen to throw an error/warn describing the source location and MethodInstance containing the dispatch,
-    as well as a stack of static dispatches that led to the dispatch (if available).
-    The option is disabled by default. (0=>disabled, 1=>error, 2=>print warnings)
+    If enabled, only provably reachable code (from functions marked with `entrypoint`) is included
+    in the output system image. Errors or warnings can be given for call sites too dynamic to handle.
+    The option is disabled by default. (0=>disabled, 1=>safe (static errors), 2=>unsafe, 3=>unsafe plus warnings)
     """
-    static_call_graph::Cint
+    trim::Cint
 
     """
     A pointer of type
@@ -1526,14 +1526,14 @@ struct CodegenParams
                    prefer_specsig::Bool=false,
                    gnu_pubnames::Bool=true, debug_info_kind::Cint = default_debug_info_kind(),
                    debug_info_level::Cint = Cint(JLOptions().debug_level), safepoint_on_entry::Bool=true,
-                   gcstack_arg::Bool=true, use_jlplt::Bool=true, static_call_graph::Cint=Cint(0),
+                   gcstack_arg::Bool=true, use_jlplt::Bool=true, trim::Cint=Cint(0),
                    lookup::Ptr{Cvoid}=unsafe_load(cglobal(:jl_rettype_inferred_addr, Ptr{Cvoid})))
         return new(
             Cint(track_allocations), Cint(code_coverage),
             Cint(prefer_specsig),
             Cint(gnu_pubnames), debug_info_kind,
             debug_info_level, Cint(safepoint_on_entry),
-            Cint(gcstack_arg), Cint(use_jlplt), Cint(static_call_graph),
+            Cint(gcstack_arg), Cint(use_jlplt), Cint(trim),
             lookup)
     end
 end