+++ src/jloptions.c
@@ -101,7 +101,7 @@ JL_DLLEXPORT void jl_init_options(void)
                         0, // permalloc_pkgimg
                         0, // heap-size-hint
                         0, // trace_compile_timing
-                        0, // static_call_graph
+                        0, // trim
     };
     jl_options_initialized = 1;
 }
@@ -266,7 +266,7 @@ static const char opts_hidden[]  =
     "                                               compile in ms\n"
     " --image-codegen                               Force generate code in imaging mode\n"
     " --permalloc-pkgimg={yes|no*}                  Copy the data section of package images into memory\n"
-    " --static-call-graph={no*|safe|unsafe|unsafe-warn}\n"
+    " --trim={no*|safe|unsafe|unsafe-warn}\n"
     "                                               Build a sysimage including only code provably reachable\n"
     "                                               from methods marked by calling `entrypoint`. In unsafe\n"
     "                                               mode, the resulting binary might be missing needed code\n"
@@ -320,7 +320,7 @@ JL_DLLEXPORT void jl_parse_opts(int *argcp, char ***argvp)
            opt_heap_size_hint,
            opt_gc_threads,
            opt_permalloc_pkgimg,
-           opt_static_call_graph,
+           opt_trim,
     };
     static const char* const shortopts = "+vhqH:e:E:L:J:C:it:p:O:g:m:";
     static const struct option longopts[] = {
@@ -384,7 +384,7 @@ JL_DLLEXPORT void jl_parse_opts(int *argcp, char ***argvp)
         { "strip-ir",        no_argument,       0, opt_strip_ir },
         { "permalloc-pkgimg",required_argument, 0, opt_permalloc_pkgimg },
         { "heap-size-hint",  required_argument, 0, opt_heap_size_hint },
-        { "static-call-graph",  optional_argument, 0, opt_static_call_graph },
+        { "trim",  optional_argument, 0, opt_trim },
         { 0, 0, 0, 0 }
     };
 
@@ -944,17 +944,17 @@ restart_switch:
             else
                 jl_errorf("julia: invalid argument to --permalloc-pkgimg={yes|no} (%s)", optarg);
             break;
-        case opt_static_call_graph:
+        case opt_trim:
             if (optarg == NULL || !strcmp(optarg,"safe"))
-                jl_options.static_call_graph = JL_STATIC_CALL_GRAPH_SAFE;
+                jl_options.trim = JL_TRIM_SAFE;
             else if (!strcmp(optarg,"no"))
-                jl_options.static_call_graph = JL_STATIC_CALL_GRAPH_NO;
+                jl_options.trim = JL_TRIM_NO;
             else if (!strcmp(optarg,"unsafe"))
-                jl_options.static_call_graph = JL_STATIC_CALL_GRAPH_UNSAFE;
+                jl_options.trim = JL_TRIM_UNSAFE;
             else if (!strcmp(optarg,"unsafe-warn"))
-                jl_options.static_call_graph = JL_STATIC_CALL_GRAPH_UNSAFE_WARN;
+                jl_options.trim = JL_TRIM_UNSAFE_WARN;
             else
-                jl_errorf("julia: invalid argument to --static-call-graph={safe|no|unsafe|unsafe-warn} (%s)", optarg);
+                jl_errorf("julia: invalid argument to --trim={safe|no|unsafe|unsafe-warn} (%s)", optarg);
             break;
         default:
             jl_errorf("julia: unhandled option -- %c\n"