+++ src/precompile_utils.c
@@ -342,7 +342,7 @@ JL_DLLEXPORT void jl_add_ccallable_entrypoints(void)
     jl_foreach_reachable_mtable(enq_ccallable_entrypoints, NULL);
 }
 
-static void *jl_precompile_static_call_graph(size_t world)
+static void *jl_precompile_trimmed(size_t world)
 {
     // array of MethodInstances and ccallable aliases to include in the output
     jl_array_t *m = jl_alloc_vec_any(0);
@@ -362,7 +362,7 @@ static void *jl_precompile_static_call_graph(size_t world)
     }
 
     jl_cgparams_t params = jl_default_cgparams;
-    params.static_call_graph = jl_options.static_call_graph;
+    params.trim = jl_options.trim;
     void *native_code = jl_create_native(m, NULL, &params, 0, /* imaging */ 1, 0,
                                          world);
     JL_GC_POP();